pyproject.toml
[project]
name = "agentic-ai-langgraph"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "arxiv>=2.4.1",
    "langchain==1.2.13",
    "langchain-community==0.4.1",
    "langchain-core==1.2.22",
    "langchain-groq>=1.1.2",
    "langchain-ollama>=1.0.1",
    "langchain-tavily>=0.2.17",
    "langgraph==1.1.3",
    "langgraph-cli[inmem]>=0.4.19",
    "python-dotenv>=1.2.2",
    "wikipedia>=1.4.0",
]
